Move `DbIterator` and `MessageDbIterator` to `query` package
