+++ rdbms/src/main/java/io/spine/server/storage/jdbc/query/DbIterator.java
@@ -18,10 +18,12 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.server.storage.jdbc;
+package io.spine.server.storage.jdbc.query;
 
 import com.google.common.annotations.VisibleForTesting;
 import io.spine.annotation.Internal;
+import io.spine.server.storage.jdbc.DatabaseException;
+import io.spine.server.storage.jdbc.Serializer;
 
 import java.io.Closeable;
 import java.sql.Connection;
