+++ rdbms/src/main/java/io/spine/server/storage/jdbc/query/MessageDbIterator.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.server.storage.jdbc;
+package io.spine.server.storage.jdbc.query;
 
 import com.google.protobuf.Message;
 import io.spine.annotation.Internal;
