+++ rdbms/src/test/java/io/spine/server/storage/jdbc/query/DbIteratorShould.java
@@ -18,8 +18,9 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.server.storage.jdbc;
+package io.spine.server.storage.jdbc.query;
 
+import io.spine.server.storage.jdbc.DatabaseException;
 import org.junit.Test;
 
 import java.sql.Connection;
